Adds hero section
The hero section is added, but far from perfect. Material UI components have been used to construct it. Right now I have added up to a Join button.

That Join buttion will be a Link (react-router-dom component) to a sign up form.

In the future I should consider adding a Login button on the Navbar for existing users