Adds PanelLeft component
While building the Dashboard there was a considerable amount of code going into the component itself.

I made the choice to modularize the elements of the utility panel at the left--hand side of the user dashboard. That way, if I want this utility panel accessible on other display, I can reuse it.