Updates Card design
Refactored the Card component. Previously it was a simple two sided card, with text only.

The card now features a CardHeader (MUI) that will show the card's topic (the subject, or deck it's in - all ways to say the same thing, I think). Only on the front of the card.

Also applies a colored banner to the top of the card. Eventually I need to find a way to programmatically make this color dynamic - allowing users to assign colors to topics and color code their studying.

Additionally, delete and edit Buttons (MUI) (non-functional) have been added to each card. Only edit on back.