Updates Dashboard flow
Some needless div were removed from the JSX. I also re-assessed how I was using the MUI components.

The render methods of the Cards and Card components have now been wrapped with a <React.Fragment> tag instead of another MUI Paper or Card - this was adding a lot of divs to the JSX that really didn't need to be there.

Finally, instead of wrapping the <Cards /> render (in Dashboard) in a Paper component - which, as per MUI docs, is intended to be a smooth, paper-like place to display text information - it is now wrapped in a Box MUI component. This component is a better choice because it's a bit of a jack-of-all-trades utility component - a highly customizable div.