Adds MirageJS
MirageJS has been added to the project to mock a RESTful API for the sake of front end development, and later for TESTING the front end without relying on an external service (i.e. - the API I'm currently building for LeitSpeed)

CardsData has now been moved into the Mirage server. Dashboard is making a useEffect fetch call, Mirage intercepts the call and routes the request to our in-built 'endpoint'