Adds Theme Context
Corrected the file name of the landing directory in the components (the directory was named with a capital, it didn't fit, so that's why you see those changes)

Implemented a custom hook to toggle a piece of state that's a boolean value.
Implemented the beginning of a Theme Context, you can't see it doing much on the homepage right now, but in login you'll see that toggle switch on the nav bar now changes the navbar light, and the background black.

Also created a view component, PageContent, to wrap the background color around children and simplify that whole thing.