Updates dark mode
Refined some of the dark mode rules.

There are notes in the NavBarStyles.js file regarding the light and dark mode.

I'm repeating myself a lot to define the dark mode rules... and I suspect there's a smarter way to do that (more DRY)

There's also a problem.... when a user selects dark mode, and then goes to another page, the isDarkMode variable is reset to false, and they will have to select it again on render.

How can I persist the isDarkMode setting over multiple renders once it's been selected??